requirements.txt
streamlit==1.47.0
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.28
python-dotenv==1.1.1
qdrant-client==1.14.3
fastembed==0.7.1
openai==1.97.0
pypdf==5.8.0